+++ packages/source-models/src/normalization/normalizeArcs.spec.ts
@@ -1,5 +1,4 @@
-import { expect } from "chai"
-import { describe, it } from "mocha"
+import { describe, expect, it } from "vitest"
 import { Arc } from "../types/Arc"
 import normalizeArcs from "./normalizeArcs"
 describe("normalizeArcs", () => {