+++ packages/utils-aws/src/errors/isAWSError.spec.ts
@@ -1,5 +1,4 @@
-import { expect } from "chai"
-import { describe, it } from "mocha"
+import { describe, expect, it } from "vitest"
 import isAWSError from "./isAWSError"
 describe("isAwSError", () => {
     it("should detect a valid error", () => {