+++ apps/api/CHANGELOG.md
@@ -11,7 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
--   Upgraded `@aws-sdk/client-s3` to `3.262.0`.
+-   Upgrades:
+    -   Node.js 20.
+    -  `@aws-sdk/client-s3` to `3.262.0`.
 
 ### Deprecated